ui: extract txn details api into its own file
This commit simply moves txn details functions out of `txnDetailsApi`
and into its own file. No functional changes are made. New filesto
organize shared functions:
- txnInsightsUtils.ts

This patch also moves contention related functions from `txnInsightsApi`
to `contentionApi`.

Epic: none

Release note: None
