+++ pkg/ui/workspaces/cluster-ui/src/api/contentionApi.ts
@@ -10,16 +10,31 @@
 
 import {
   executeInternalSql,
+  formatApiResult,
+  INTERNAL_SQL_API_APP,
   LARGE_RESULT_SIZE,
   LONG_TIMEOUT,
+  sqlApiErrorMessage,
   SqlApiResponse,
   SqlExecutionRequest,
+  SqlExecutionResponse,
   sqlResultsAreEmpty,
-  formatApiResult,
 } from "./sqlApi";
-import { ContentionDetails } from "src/insights";
+import {
+  ContentionDetails,
+  InsightExecEnum,
+  InsightNameEnum,
+  TxnContentionInsightDetails,
+} from "src/insights";
 import moment from "moment-timezone";
-import { getLogger } from "../util";
+import { FixFingerprintHexValue, getLogger } from "../util";
+import { TxnInsightDetailsRequest } from "./txnInsightDetailsApi";
+import { makeInsightsSqlRequest } from "./txnInsightsUtils";
+import {
+  FingerprintStmtsResponseColumns,
+  TxnStmtFingerprintsResponseColumns,
+  TxnWithStmtFingerprints,
+} from "./txnInsightsApi";
 
 export type ContentionFilters = {
   waitingTxnID?: string;
@@ -172,12 +187,216 @@ function contentionDetailsQuery(filters?: ContentionFilters) {
                     encode(waiting_stmt_fingerprint_id, 'hex') AS waiting_stmt_fingerprint_id,
                     contention_duration,
                     contending_pretty_key AS key,
-  database_name,
-  schema_name,
-  table_name,
-  index_name
+                    database_name,
+                    schema_name,
+                    table_name,
+                    index_name
     FROM
-      crdb_internal.transaction_contention_events AS tce
+      crdb_internal.transaction_contention_events
       ${whereClause}
   `;
+}
+
+type PartialTxnContentionDetails = Omit<
+  TxnContentionInsightDetails,
+  "application" | "queries"
+>;
+
+function formatTxnContentionDetailsResponse(
+  response: ContentionDetails[],
+): PartialTxnContentionDetails {
+  if (!response || response.length === 9) {
+    // No data.
+    return;
+  }
+
+  const row = response[0];
+  return {
+    transactionExecutionID: row.waitingTxnID,
+    transactionFingerprintID: FixFingerprintHexValue(
+      row.waitingTxnFingerprintID,
+    ),
+    blockingContentionDetails: response,
+    insightName: InsightNameEnum.highContention,
+    execType: InsightExecEnum.TRANSACTION,
+  };
+}
+
+function buildTxnContentionInsightDetails(
+  partialTxnContentionDetails: PartialTxnContentionDetails,
+  txnsWithStmtFingerprints: TxnWithStmtFingerprints[],
+  stmtFingerprintToQuery: StmtFingerprintToQueryRecord,
+): TxnContentionInsightDetails {
+  if (!partialTxnContentionDetails && !txnsWithStmtFingerprints.length) {
+    return null;
+  }
+
+  partialTxnContentionDetails.blockingContentionDetails.forEach(blockedRow => {
+    const currBlockedFingerprintStmts = txnsWithStmtFingerprints.find(
+      txn =>
+        txn.transactionFingerprintID === blockedRow.blockingTxnFingerprintID,
+    );
+
+    if (!currBlockedFingerprintStmts) {
+      return;
+    }
+
+    blockedRow.blockingTxnQuery = currBlockedFingerprintStmts.queryIDs.map(
+      id =>
+        stmtFingerprintToQuery.get(id) ??
+        `Query unavailable for statement fingerprint ${id}`,
+    );
+  });
+
+  const waitingTxn = txnsWithStmtFingerprints.find(
+    txn =>
+      txn.transactionFingerprintID ===
+      partialTxnContentionDetails.transactionFingerprintID,
+  );
+
+  return {
+    ...partialTxnContentionDetails,
+    application: waitingTxn?.application,
+  };
+}
+
+type StmtFingerprintToQueryRecord = Map<
+  string, // Key = Stmt fingerprint ID
+  string // Value = query string
+>;
+
+function createStmtFingerprintToQueryMap(
+  response: SqlExecutionResponse<FingerprintStmtsResponseColumns>,
+): StmtFingerprintToQueryRecord {
+  const idToQuery: Map<string, string> = new Map();
+  if (!response || sqlResultsAreEmpty(response)) {
+    // No statement fingerprint results.
+    return idToQuery;
+  }
+  response.execution.txn_results[0].rows.forEach(row => {
+    idToQuery.set(
+      FixFingerprintHexValue(row.statement_fingerprint_id),
+      row.query,
+    );
+  });
+
+  return idToQuery;
+}
+
+// Query to select all statement queries for each requested statement
+// fingerprint.
+const fingerprintStmtsQuery = (stmtFingerprintIDs: string[]): string => `
+SELECT
+  DISTINCT ON (fingerprint_id) encode(fingerprint_id, 'hex') AS statement_fingerprint_id,
+  (metadata ->> 'query') AS query
+FROM crdb_internal.statement_statistics_persisted
+WHERE encode(fingerprint_id, 'hex') =
+      ANY ARRAY[ ${stmtFingerprintIDs.map(id => `'${id}'`).join(",")} ]`;
+
+// txnStmtFingerprintsQuery selects all statement fingerprints for each
+// requested transaction fingerprint.
+const txnStmtFingerprintsQuery = (txnFingerprintIDs: string[]) => `
+SELECT
+  DISTINCT ON (fingerprint_id) encode(fingerprint_id, 'hex') AS transaction_fingerprint_id,
+  app_name,
+  ARRAY( SELECT jsonb_array_elements_text(metadata -> 'stmtFingerprintIDs' )) AS query_ids
+FROM crdb_internal.transaction_statistics_persisted
+WHERE app_name != '${INTERNAL_SQL_API_APP}'
+  AND encode(fingerprint_id, 'hex') = 
+      ANY ARRAY[ ${txnFingerprintIDs.map(id => `'${id}'`).join(",")} ]`;
+
+function formatTxnFingerprintsResults(
+  response: SqlExecutionResponse<TxnStmtFingerprintsResponseColumns>,
+): TxnWithStmtFingerprints[] {
+  if (sqlResultsAreEmpty(response)) {
+    return [];
+  }
+
+  return response.execution.txn_results[0].rows.map(row => ({
+    transactionFingerprintID: FixFingerprintHexValue(
+      row.transaction_fingerprint_id,
+    ),
+    queryIDs: row.query_ids.map(id => FixFingerprintHexValue(id)),
+    application: row.app_name,
+  }));
+}
+
+export async function getTxnInsightsContentionDetailsApi(
+  req: Pick<TxnInsightDetailsRequest, "txnExecutionID">,
+): Promise<TxnContentionInsightDetails> {
+  // Note that any errors encountered fetching these results are caught
+  // earlier in the call stack.
+  //
+  // There are 3 api requests/queries in this process.
+  // 1. Get contention insight for the requested transaction.
+  // 2. Get the stmt fingerprints for ALL transactions involved in the contention.
+  // 3. Get the query strings for ALL statements involved in the transaction.
+
+  // Get contention results for requested transaction.
+
+  const contentionResponse = await getContentionDetailsApi({
+    waitingTxnID: req.txnExecutionID,
+  });
+  const contentionResults = contentionResponse.results;
+
+  if (contentionResults.length === 0) {
+    return null;
+  }
+
+  const contentionDetails =
+    formatTxnContentionDetailsResponse(contentionResults);
+
+  // Collect all blocking txn fingerprints involved.
+  const txnFingerprintIDs: string[] = [];
+  contentionDetails.blockingContentionDetails.forEach(x =>
+    txnFingerprintIDs.push(x.blockingTxnFingerprintID),
+  );
+
+  // Request all blocking stmt fingerprint ids involved.
+  const getStmtFingerprintsResponse =
+    await executeInternalSql<TxnStmtFingerprintsResponseColumns>(
+      makeInsightsSqlRequest([txnStmtFingerprintsQuery(txnFingerprintIDs)]),
+    );
+  if (getStmtFingerprintsResponse.error) {
+    throw new Error(
+      `Error while retrieving statements information: ${sqlApiErrorMessage(
+        getStmtFingerprintsResponse.error.message,
+      )}`,
+    );
+  }
+
+  const txnsWithStmtFingerprints = formatTxnFingerprintsResults(
+    getStmtFingerprintsResponse,
+  );
+
+  const stmtFingerprintIDs = new Set<string>();
+  txnsWithStmtFingerprints.forEach(txnFingerprint =>
+    txnFingerprint.queryIDs.forEach(id => stmtFingerprintIDs.add(id)),
+  );
+
+  // Request query string from stmt fingerprint ids.
+  let stmtQueriesResponse: SqlExecutionResponse<FingerprintStmtsResponseColumns> | null =
+    null;
+
+  if (stmtFingerprintIDs.size) {
+    stmtQueriesResponse =
+      await executeInternalSql<FingerprintStmtsResponseColumns>(
+        makeInsightsSqlRequest([
+          fingerprintStmtsQuery(Array.from(stmtFingerprintIDs)),
+        ]),
+      );
+    if (stmtQueriesResponse.error) {
+      throw new Error(
+        `Error while retrieving statements information: ${sqlApiErrorMessage(
+          stmtQueriesResponse.error.message,
+        )}`,
+      );
+    }
+  }
+
+  return buildTxnContentionInsightDetails(
+    contentionDetails,
+    txnsWithStmtFingerprints,
+    createStmtFingerprintToQueryMap(stmtQueriesResponse),
+  );
 }
