+++ pkg/ui/workspaces/cluster-ui/src/api/index.ts
@@ -27,3 +27,4 @@ export * from "./databaseDetailsApi";
 export * from "./tableDetailsApi";
 export * from "./types";
 export * from "./jobProfilerApi";
+export * from "./txnInsightDetailsApi";
