+++ pkg/ui/workspaces/cluster-ui/src/insights/types.ts
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-import { Moment } from "moment-timezone";
+import moment, { Moment } from "moment-timezone";
 import { Filters } from "../queryFilter";
 
 // This enum corresponds to the string enum for `problems` in `cluster_execution_insights`
@@ -85,6 +85,7 @@ export type ContentionDetails = {
   contentionTimeMs: number;
 };
 
+// The return type of getTxnInsightsContentionDetailsApi.
 export type TxnContentionInsightDetails = {
   transactionExecutionID: string;
   application: string;
@@ -101,7 +102,6 @@ export type TxnInsightDetails = {
   txnDetails?: TxnInsightEvent;
   blockingContentionDetails?: ContentionDetails[];
   statements?: StmtInsightEvent[];
-  execType?: InsightExecEnum;
 };
 
 // Shown on the stmt insights overview page.
